+++ hazelcast-documentation/src/WhatsNew.md
@@ -7,9 +7,10 @@
 ### New Features
 This section provides the new features introduced with Hazelcast 3.4 release. 
 
-- Back Pressure - ??? Please see [Back Pressure](#back-pressure).
-- High Density for JCache - Please see [High Density for JCache](#high-density-for-jcache).
-- Jetty Specific Session Replication - Please see [Jetty Based Web Session Replication](#jetty-based-web-session-replication).
+- High Density for JCache: ??? Please see [High Density for JCache](#high-density-for-jcache).
+- Back Pressure: ??? Please see [Back Pressure](#back-pressure).
+- Jetty Specific Session Replication: ??? Please see [Jetty Based Web Session Replication](#jetty-based-web-session-replication).
+- Hazelcast Configuration Import: ???